[ci]Rename WindowStylesReportTool to StylesReportTool (#17950)
