+++ .pipelines/ci/templates/build-powertoys-steps.yml
@@ -77,18 +77,18 @@ steps:
     maximumCpuCount: true
 
 - task: NuGetCommand@2
-  displayName: Restore NuGet packages for WindowStylesReportTool.sln
+  displayName: Restore NuGet packages for StylesReportTool.sln
   inputs:
     command: restore
     feedsToUse: config
     configPath: NuGet.config
-    restoreSolution: tools\WindowStylesReportTool\WindowStylesReportTool.sln
-    restoreDirectory: '$(Build.SourcesDirectory)\tools\WindowStylesReportTool\packages'
+    restoreSolution: tools\StylesReportTool\StylesReportTool.sln
+    restoreDirectory: '$(Build.SourcesDirectory)\tools\StylesReportTool\packages'
 
 - task: VSBuild@1
-  displayName: 'Build WindowStylesReportTool.sln'
+  displayName: 'Build StylesReportTool.sln'
   inputs:
-    solution: '**\WindowStylesReportTool.sln'
+    solution: '**\StylesReportTool.sln'
     vsVersion: 17.0
     platform: '$(BuildPlatform)'
     configuration: '$(BuildConfiguration)'
