+++ installer/PowerToysSetup/Product.wxs
@@ -616,8 +616,8 @@
             <Component Id="WebcamReportTool_exe" Win64="yes">
                 <File Source="$(var.BinDir)WebcamReportTool\PowerToys.WebcamReportTool.exe" Id="WebcamReportTool.exe" Checksum="yes" />
             </Component>
-			<Component Id="WindowStylesReportTool_exe" Win64="yes">
-				<File Source="$(var.BinDir)WindowStylesReportTool\PowerToys.WindowStylesReportTool.exe" Id="WindowStylesReportTool.exe" Checksum="yes" />
+			<Component Id="StylesReportTool_exe" Win64="yes">
+				<File Source="$(var.BinDir)StylesReportTool\PowerToys.StylesReportTool.exe" Id="StylesReportTool.exe" Checksum="yes" />
 			</Component>
         </DirectoryRef>
 
@@ -1052,7 +1052,7 @@
         <ComponentGroup Id="ToolComponents" Directory="ToolsFolder">
             <ComponentRef Id="BugReportTool_exe" />
             <ComponentRef Id="WebcamReportTool_exe" />
-			<ComponentRef Id="WindowStylesReportTool_exe" />
+			<ComponentRef Id="StylesReportTool_exe" />
         </ComponentGroup>
     </Fragment>
